[KRITIMA] Hydro class modified

+++ C0778923_W2020_MID_TERM_MAD3004/C0778923_W2020_MID_TERM_MAD3004/Hydro.swift
@@ -10,9 +10,9 @@ import Foundation
 
 class Hydro :Bill {
     var agencyName: String
-    var unitConsumed: Float
+    var unitConsumed: Int
     
- init(aName: String, uConsumed: Float) {
+ init(aName: String, uConsumed: Int) {
       self.agencyName = aName
       self.unitConsumed = uConsumed
   }
